# Conflicts:
#	.settings/org.eclipse.core.resources.prefs
#	WebContent/admin/memberMng.jsp
#	WebContent/main/boardAsk.jsp
#	src/handler/user/AskCompProHandler.java

+++ WebContent/survey/uploadSur.jsp
@@ -16,45 +16,23 @@
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
 <!DOCTYPE html>
 
-
-<!-- 자바스크립트 -->
-
-<!-- JQuery -->
-<!-- <script type="text/javascript" src="/DotNet/jquery-3.4.1.js"></script> -->
-<!-- <script
-  src="https://code.jquery.com/jquery-3.4.1.js"
-  integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU="
-  crossorigin="anonymous"></script>
-
-BootStrap cdn 인터넷에서 끌어 사용
-<link rel="stylesheet" href="//unpkg.com/bootstrap@4/dist/css/bootstrap.min.css">
-<script src='//unpkg.com/jquery@3/dist/jquery.min.js'></script>
-<script src='//unpkg.com/popper.js@1/dist/umd/popper.min.js'></script>
-<script src='//unpkg.com/bootstrap@4/dist/js/bootstrap.min.js'></script> -->
-
-
-
-<!-- 상단 바 -->
-
-<!-- 	<nav class="navbar navbar-expand-md navbar-default"> -->
-<%-- 		<jsp:include page="../mainNav.do" flush="false"></jsp:include> --%>
-<!-- 	</nav> -->
-<!-- 상단 바  끝-->
 <html>
 <head>
 <meta charset="UTF-8">
 <meta name="viewport" content="width=device-width, initial-scale=1.0">
 <title>결과</title>
-<link rel="stylesheet" href="bootstrap.css">
+<link rel="stylesheet" href="/DotNet/css/bootstrap.css">
 <link rel="stylesheet"
 	href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/css/bootstrap.min.css">
 <link
 	href="https://fonts.googleapis.com/css?family=Caveat|Concert+One|Dancing+Script|Fredoka+One|Kalam|Kaushan+Script|Lobster|Luckiest+Guy|Merienda|Neucha|Sniglet|ZCOOL+QingKe+HuangYou&display=swap"
 	rel="stylesheet">
 <!-- style_dotnet.csa -->
-<link rel="stylesheet" type="text/css" href="style_dotnet.css" />
-<script type="text/javascript" src="jquery-3.4.1.js"></script>
-<script type="text/javascript" src="bootstrap.bundle.js"></script>
+
+<link rel="stylesheet" type="text/css" href="/DotNet/css/style_dotnet.css" />
+<script type="text/javascript" src="/DotNet/js/jquery-3.4.1.js"></script>
+<script type="text/javascript" src="/DotNet/js/bootstrap.bundle.js"></script>
+
 
 
 <style>
@@ -102,7 +80,7 @@ img {
 </style>
 </head>
 <body>
-<h1 align="center">설문번호 : ${two.s_num}</h1>
+<h1 align="center" id="${two.s_num}">설문번호 : ${two.s_num}</h1>
 <br>
 <!-- 추가~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ id = intro -->
 <div class="container-fluid mt-5" id="intro">
@@ -111,7 +89,7 @@ img {
 			<label class="col-sm-2 col-form-label text-right">이름</label>
 			<div class="col-sm-8">
 				<c:if test="${sessionScope.memId eq null}">
-					<input type="text" class="form-control"  name="name"
+					<input type="text" class="form-control"  name="name" id="noMem"
 						placeholder="이름을 입력해주세요" form="choiceInfo" ><!--    -->
 				</c:if>
 				<c:if test="${sessionScope.memId ne null}">
@@ -148,10 +126,10 @@ img {
 	</c:forEach>
 	
 	<div class="bottom">
-		<input type="submit" value="완료" hidden>
+		<input type="button" value="완료" id="cbtn" hidden>
 		<form action="uploadResult.do" method="post" name="choiceInfo" id="choiceInfo">
 			<c:forEach var="ii" begin="${1}" end="${size}">
-				<input type="hidden" name="q${ii}">
+				<input type="hidden" name="q${ii}" class="qList" >
 			</c:forEach>
 			<input type="hidden" name="s_num" value="${two.s_num}"> <input
 				type="hidden" name="point" value="${boardDto.point }">
@@ -160,6 +138,9 @@ img {
 		<a href="main.do" class="tomain">메인으로</a>
 	</div>
 </div>
+<script type="text/javascript" src="jquery-3.4.1.js"></script>
+<script type="text/javascript" src="bootstrap.bundle.js"></script>
+
 <script>
 /////////////추가!!!!!!!!!?//////////////////
 $(document).ready(function(){
@@ -217,10 +198,6 @@ $(document).ready(function(){
 	}
 	$(document).ready(function() {
 
-		// 		$.ajax({
-
-		// 			data:
-		// 		})
 		show(0, 1, 0);
 		img.on('click', function() {
 			img.not($(this)).animate({
@@ -236,14 +213,33 @@ $(document).ready(function(){
 // 			console.log(i1 + " , " + i2 + "  , " + qn)
 			show(i1, i2, qn);
 			if (qn == qlen) {
+				var snum = $('h1').prop('id')
+				var data = { cnt : clickCnt , s_num : snum, type : "2" , id :$('#noMem').val()};
+				$.ajax({
+					url:"clickLog.do",
+					type:"post",
+					data:data,
+					dataType:"text",
+					success:function(data){
+						/// 광클시 에러 --> 처음으로 되돌리기
+						
+					}
+													
+				});
 				setTimeout(function() {
-					$("input[type=submit]").click()
+					$("#cbtn").click();
 				}, 2000);
 			}
 		});
 
 	});
 </script>
+		<script>
+			var clickCnt = 0;
+			$('body, html').on('click',function(){
+				clickCnt++;
+			});
+		</script>
 <script>
 	var q_num = 0;
 	$(document).ready(
@@ -262,22 +258,41 @@ $(document).ready(function(){
 					// 선택정보 저장
 					localStorage.setItem("q" + q_num, $(this).attr("name"));
 				});
-				$("input:submit").click(
+				$("#cbtn").click(
 						function() {
 							var params = document.choiceInfo;
-							var size = ${size}
+							var size = "${size}"
 // 							alert("size = " + size);
 							for (var i = 1; i <= size; i++) {
-								$("input[name=q" + i + "]").attr("value",
-										localStorage.getItem("q" + i));
-// 								alert("q"+i+"="+localStorage.getItem("q"+i)); 
+								$("input[name=q" + i + "]").attr("value", localStorage.getItem("q" + i));
+								
 							}
-
-							params.submit();
-							localStorage.clear();
+							if($('.qList').length != localStorage.length){
+								alert("클릭를 조금 교양있게 해주셔야 반영이 됩니다 ^^");
+								var result = confirm('다시 하시겠습니까??\n ok(확인)->처음, cancel(취소)->메인');
+								if(result){
+									localStorage.clear();
+									location.reload();
+								}else{
+									localStorage.clear();
+									location.href = "main.do"
+								}
+							}else{
+								params.submit();
+								localStorage.clear();
+							}
 						});
+
+				// 메인으로 가는 경우
 				$(".tomain").on("click", function() {
 					localStorage.clear();
+				});
+				// 설문 취소를 누를 경우
+				jQuery("input:reset").click(function(){
+					if(confirm("설문을 그만하시겠습니까?")){
+						localStorage.clear();
+						history.back();
+					}
 				});
 			});
 </script>
